+++ layer-atlas/src/main/java/com/layer/atlas/messagetypes/AttachmentSender.java
@@ -2,7 +2,13 @@ package com.layer.atlas.messagetypes;
 
 import android.app.Activity;
 import android.content.Intent;
+import android.content.pm.PackageManager;
 import android.os.Parcelable;
+import android.support.annotation.NonNull;
+import android.text.TextUtils;
+
+import static android.support.v4.app.ActivityCompat.requestPermissions;
+import static android.support.v4.content.ContextCompat.checkSelfPermission;
 
 /**
  * AttachmentSenders populate the AtlasMessageComposer attachment menu and handle message sending
@@ -84,5 +90,26 @@ public abstract class AttachmentSender extends MessageSender {
      */
     public Integer getIcon() {
         return mIcon;
+    }
+
+
+    /**
+     *
+     * Convenience method to check if a set of permissions have been granted
+     */
+    protected boolean hasPermissions(@NonNull Activity activity, String... permissions) {
+        for (String permission: permissions) {
+            if (TextUtils.isEmpty(permission)) continue;
+
+            if (checkSelfPermission(activity, permission) != PackageManager.PERMISSION_GRANTED) {
+                return false;
+            }
+        }
+
+        return true;
+    }
+
+    protected void requestPermissions (@NonNull Activity activity, final int permissionsCode, String... permissions) {
+        android.support.v4.app.ActivityCompat.requestPermissions(activity, permissions, permissionsCode);
     }
 }